Added Status and Gitpull Command to System Cog
